backend-django/requirements.txt
Django>=4.2
google-generativeai>=0.5.0
python-dotenv>=1.0.1
djangorestframework
gunicorn
django-cors-headers
djangorestframework-simplejwt